finished signup as staffs in menu

+++ Hospital System/src/main/java/Controllers/LoginSignUp/CheckStaffType.java
@@ -4,7 +4,7 @@ import UseCases.Staff.StaffManager;
 import UseCases.Staff.StaffManaging;
 
 public class CheckStaffType {
-    public String checkType(int id) {
+    public String checkType(long id) {
         StaffManaging sm = new StaffManager();
         return sm.getStaffType(id);
     }
